nixos-systool: added `check` command
I want to make sure I keep my systems mostly up to date, so I want an
easy way to check if the flake lock file is getting outdated.

This new command checks the timestamps from the flake lock and prints a
message based on how much time has elapsed. It's hard-coded to two weeks
for now, but I may add support for configuring it.