nixos-systool: specify flake path to nixos-rebuild
Currently, my system flake is dependent on a symlink from
`/etc/nixos/flake.nix` to the location of the system flake, but
`nixos-rebuild` will take a `--flake` option to specify the URI
directly.

Since all the subcommands in the nixos-systool take the flake path now,
it's simple to just pass this to the `nixos-rebuild` command invocation,
eliminating the need for the symlink.