nixos-systool: improve configuration and check
The Check command currently just looks at the lock file in the system
flake repo. This can be misleading because updating that lock file
doesn't mean that the configuration has been applied to the system.
Instead we now try to look at the flake that was stored in the Nix store
when the `/run/current-system` was built.

Also fixed up the configuration story. The `confy` library was too
opinionated and broke on an invalid configuration. The `figment` library
seems like a much more robust solution.