Add cargo-deny to at least try to check licenses
Using GPL3 for this tool means I need to be a little more careful about
what I include from the dependencies side of things.

This might not be needed if they're just libraries, but I haven't
researched this yet.

This also has the added benefit of checking for advisories and other
issues with dependencies.